Criando respostas padronizadas para usuario e requisições!

+++ src/main/java/com/project/ecommerce/entity/Categoria.java
@@ -20,7 +20,7 @@ public class Categoria {
     private String nome;
 
     @Version
-    private Integer versao;
+    private Long versao;
 
     @OneToMany(mappedBy = "categoria")
     private List<ProdutoEntity> produtos;
